define business models

+++ app/src/main/java/com/ayan/boxboxassignmnet/data/dto/RaceDetailsDto.kt
@@ -12,10 +12,10 @@ data class Schedule(
     val circuitId: String,
     val isSprint: Boolean,
     val podium: List<String>,
-    val raceEndTime: Int,
+    val raceEndTime: Long,
     val raceId: String,
     val raceName: String,
-    val raceStartTime: Int,
+    val raceStartTime: Long,
     val raceState: String,
     val round: Int,
     val sessions: List<Session>
@@ -24,10 +24,10 @@ data class Schedule(
 @JsonClass(generateAdapter = true)
 data class Session(
     val _id: String,
-    val endTime: Int,
+    val endTime: Long,
     val sessionId: String,
     val sessionName: String,
     val sessionState: String,
     val sessionType: String,
-    val startTime: Int
+    val startTime: Long
 )
